pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lil-bro-local"
version = "0.1.0"
description = "Local-model coding TUI — wraps Ollama for offline AI-assisted development"
requires-python = ">=3.11"
dependencies = [
    "textual>=0.80.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "pillow>=10.0",
]

[project.scripts]
lilbro-local = "src_local.app:main"

[tool.hatch.build.targets.wheel]
packages = ["src_local"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100
